+++ packages/template-retail-react-app/app/components/drawer-menu/drawer-menu.jsx
@@ -418,4 +418,4 @@ DrawerMenu.propTypes = {
     itemComponent: PropTypes.func
 }
 
-export default DrawerMenu
+export {DrawerMenu}
